API設計書に基づくユニットテスト修正(もくもく会検索部分)
